+++ Stream_support/include/CGAL/IO/File_header_OFF_impl.h
@@ -393,13 +393,15 @@ std::istream& operator>>( std::istream& in, File_header_OFF& h) {
         // facets and we do not know the genus of the surface.
         // So we add 12 and a factor of 5 percent.
         if (    h.size_of_halfedges() == 0
-             || h.size_of_halfedges() > (h.size_of_vertices()
+             || h.size_of_halfedges() > double(h.size_of_vertices()
                 + h.size_of_facets() - 2 + 12) * 2.1
              || h.size_of_halfedges() < (h.size_of_vertices()
                 + h.size_of_facets() - 2) * 2
         )
-            h.set_halfedges( int((h.size_of_vertices() +
-                                  h.size_of_facets() - 2 + 12) * 2.1));
+            h.set_halfedges( int( double(h.size_of_vertices() +
+                                         h.size_of_facets() - 2 + 12) * 2.1
+                                 )
+                            );
     }
     h.set_off_header( h.off());
     return in;
